Tối ưu lại. Mọi người chạy lại cơ sở dữ liệu trong bài nhá!

+++ QLKFC/QuanLyThucDon.cs
@@ -22,8 +22,9 @@ namespace QLKFC
         }
 
         QLBHKFCContext db = new QLBHKFCContext();
-        string filename;
-        //Dọn dữ liệu nhập
+        string filename, iname;
+
+        #region Hàm khai báo
         private void clearTextBox()
         {
             txtFind.Clear();
@@ -35,7 +36,7 @@ namespace QLKFC
             cmbLoai.SelectedItem = null;
             pcbMoTa.Image = null;
         }
-        //Lấy danh sách loại sản phẩm
+
         private void loadCmb()
         {
             try
@@ -57,30 +58,10 @@ namespace QLKFC
             }
             
         }
-        //Định dạng bảng
-        private void dinhDangBang()
-        {
-            //Đặt tiêu đề cột
-            dgv_DSSP.Columns["MaSp"].HeaderText = "Mã sản phẩm";
-            dgv_DSSP.Columns["TenSp"].HeaderText = "Tên sản phẩm";
-            dgv_DSSP.Columns["DonGia"].HeaderText = "Đơn giá";
-            dgv_DSSP.Columns["TenLsp"].HeaderText = "Tên loại sản phẩm";
-            dgv_DSSP.Columns["Loai"].HeaderText = "Loại";
-            dgv_DSSP.Columns["HinhAnh"].HeaderText = "Hình ảnh";
 
-            //Đặt kích thước cột
-            dgv_DSSP.Columns["MaSp"].Width = 70;
-            dgv_DSSP.Columns["TenSp"].AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill;
-            dgv_DSSP.Columns["DonGia"].Width = 100;
-            dgv_DSSP.Columns["TenLsp"].Width = 150;
-            dgv_DSSP.Columns["Loai"].Width = 120;
-            dgv_DSSP.Columns["HinhAnh"].AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill;
-
-            dgv_DSSP.Columns["DonGia"].DefaultCellStyle.Format = "N0";
-        }
-        //Lấy danh sách sản phẩm
         private void loadDGV()
         {
+            dgv_DSSP.Rows.Clear();
             var query = from sp in db.SanPhams
                         join lsp in db.LoaiSanPhams on sp.MaLsp equals lsp.MaLsp
                         select new
@@ -92,10 +73,15 @@ namespace QLKFC
                             sp.Loai,
                             sp.HinhAnh
                         };
-            dgv_DSSP.DataSource = query.ToList();
-            dinhDangBang();
+            foreach (var item in query)
+            {
+                if (item.HinhAnh != null)
+                    dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia, new Bitmap(pathImage() + item.HinhAnh));
+                else
+                    dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia);
+            }
         }
-        // Check lỗi bỏ trống dữ liệu
+
         private bool checkBoTrong()
         {
             if (txtGiaBan.Text == "" || txtLoai.Text == "" || txtTenMon.Text == "" || cmbLoai.SelectedItem == null)
@@ -105,14 +91,35 @@ namespace QLKFC
             return true;
         }
 
+        private string pathImage()
+        {
+            //Lấy đường dẫn thư mục lưu ảnh
+            string pathProject = Application.StartupPath;
+            string newPath = pathProject.Substring(0, pathProject.Length - 25) + "Image" + '\\';
+            return newPath;
+        }
+        #endregion
+
         #region Tương tác người dùng
         private void btnChonAnh_Click(object sender, EventArgs e)
         {
-            OpenFileDialog browse = new OpenFileDialog() { Filter = "All files|*.*|Pictures files (*.jpg, *.jpeg, *.jpe, *.jfif, *.png)|*.jpg; *.jpeg; *.jpe; *.jfif; *.png|Png files(*.png)|*.png|Jpeg files(*.jpeg)|*.jpeg|Jpe files(*.jpe)|*.jpe|Jpg files(*.jpg)|*.jpg" };
-            if (browse.ShowDialog() == DialogResult.OK)
+            try
             {
-                filename = browse.FileName;
-                pcbMoTa.Image = Image.FromFile(filename);
+                OpenFileDialog browse = new OpenFileDialog()
+                {
+                    Title = "Chọn ảnh",
+                    Filter = "All files|*.*|Pictures files (*.jpg, *.jpeg, *.jpe, *.jfif, *.png)|*.jpg; *.jpeg; *.jpe; *.jfif; *.png|Png files(*.png)|*.png|Jpeg files(*.jpeg)|*.jpeg|Jpe files(*.jpe)|*.jpe|Jpg files(*.jpg)|*.jpg"
+                };
+                if (browse.ShowDialog() == DialogResult.OK)
+                {
+                    iname = Path.GetFileName(browse.FileName);
+                    filename = browse.FileName;
+                    pcbMoTa.Image = Image.FromFile(browse.FileName);
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message, "Thông báo lỗi", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
         }
         private void dgv_DSSP_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -128,13 +135,7 @@ namespace QLKFC
                 cmbLoai.Text = db.LoaiSanPhams.SingleOrDefault(s => s.MaLsp.Equals(sp.MaLsp)).TenLsp;
                 try
                 {
-                    MemoryStream mstr = new MemoryStream(sp.HinhAnh.ToArray());
-                    Image img = Image.FromStream(mstr);
-                    if (img == null)
-                    {
-                        return;
-                    }
-                    pcbMoTa.Image = img;
+                    pcbMoTa.Image = new Bitmap(pathImage() + sp.HinhAnh);
                 }
                 catch (Exception)
                 {
@@ -153,23 +154,45 @@ namespace QLKFC
                     SanPham sp = new SanPham();
                     sp.TenSp = txtTenMon.Text;
                     sp.MaLsp = db.LoaiSanPhams.SingleOrDefault(s => s.TenLsp.Equals(cmbLoai.SelectedItem.ToString())).MaLsp;
-                    sp.DonGia = int.Parse(txtGiaBan.Text);
-                    sp.Loai = txtLoai.Text;
-                    sp.Mota = txtMoTa.Text;
-                    MemoryStream str = new MemoryStream();
-                    pcbMoTa.Image.Save(str, System.Drawing.Imaging.ImageFormat.Jpeg);
-                    sp.HinhAnh = str.ToArray();
                     try
                     {
-                        db.SanPhams.Add(sp);
-                        db.SaveChanges();
-                        loadDGV();
-                        clearTextBox();
+                        if (int.Parse(txtGiaBan.Text)<1)
+                        {
+                            errorProvider1.SetError(txtGiaBan, "Giá bán bải là số dương!");
+                            txtGiaBan.Focus();
+                            txtGiaBan.SelectAll();
+                        }
+                        else
+                        {
+                            sp.DonGia = int.Parse(txtGiaBan.Text);
+                            sp.Loai = txtLoai.Text;
+                            sp.Mota = txtMoTa.Text;
+                            if (iname != null)
+                            {
+                                sp.HinhAnh = iname;
+                                if (!File.Exists(pathImage() + iname))
+                                {
+                                    File.Copy(filename, pathImage() + iname);
+                                }
+                            }
+                            try
+                            {
+                                db.SanPhams.Add(sp);
+                                db.SaveChanges();
+                                iname = null;
+                                MessageBox.Show("Đã thêm sản phẩm");
+                            }
+                            catch (Exception ex)
+                            {
+                                MessageBox.Show(ex.Message.ToString(), "Thông báo");
+                            }
+                            loadDGV();
+                            clearTextBox();
+                        }
                     }
-                    catch (Exception ex)
+                    catch
                     {
-
-                        MessageBox.Show(ex.Message.ToString(), "Thông báo");
+                        errorProvider1.SetError(txtGiaBan, "Phải nhập giá tiền là số!");
                     }
                 }
                 else
@@ -185,18 +208,55 @@ namespace QLKFC
 
         private void btnSua_Click(object sender, EventArgs e)
         {
-            var spSua = db.SanPhams.SingleOrDefault(sp => sp.MaSp == int.Parse(txtMaMon.Text));
-            spSua.TenSp = txtTenMon.Text;
-            spSua.MaLsp = db.LoaiSanPhams.SingleOrDefault(s => s.TenLsp.Equals(cmbLoai.SelectedItem.ToString())).MaLsp;
-            spSua.DonGia = int.Parse(txtGiaBan.Text);
-            spSua.Loai = txtLoai.Text;
-            spSua.Mota = txtMoTa.Text;
-            MemoryStream str = new MemoryStream();
-            pcbMoTa.Image.Save(str, System.Drawing.Imaging.ImageFormat.Jpeg);
-            spSua.HinhAnh = str.ToArray();
-            db.SaveChanges();
-            loadDGV();
-            clearTextBox();
+            if (checkBoTrong())
+            {
+                var spSua = db.SanPhams.SingleOrDefault(sp => sp.MaSp == int.Parse(txtMaMon.Text));
+                spSua.TenSp = txtTenMon.Text;
+                spSua.MaLsp = db.LoaiSanPhams.SingleOrDefault(s => s.TenLsp.Equals(cmbLoai.SelectedItem.ToString())).MaLsp;
+                try
+                {
+                    if (int.Parse(txtGiaBan.Text) < 1)
+                    {
+                        errorProvider1.SetError(txtGiaBan, "Giá bán bải là số dương!");
+                        txtGiaBan.Focus();
+                        txtGiaBan.SelectAll();
+                    }
+                    else
+                    {
+                        spSua.DonGia = int.Parse(txtGiaBan.Text);
+                        spSua.Loai = txtLoai.Text;
+                        spSua.Mota = txtMoTa.Text;
+                        if (iname != null)
+                        {
+                            spSua.HinhAnh = iname;
+                            if (!File.Exists(pathImage() + iname))
+                            {
+                                File.Copy(filename, pathImage() + iname);
+                            }
+                        }
+                        try
+                        {
+                            db.SaveChanges();
+                            loadDGV();
+                            clearTextBox();
+                            iname = null;
+                            MessageBox.Show("Đã được sửa");
+                        }
+                        catch (Exception ex)
+                        {
+                            MessageBox.Show(ex.Message.ToString(), "Thông báo");
+                        }
+                    }
+                }
+                catch
+                {
+                    errorProvider1.SetError(txtGiaBan, "Phải nhập giá tiền là số!");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Đã tồn tại loại sản phẩm này!", "Thông báo");
+            }
         }
 
         private void btnHuyBo_Click(object sender, EventArgs e)
@@ -204,31 +264,40 @@ namespace QLKFC
             clearTextBox();
         }
 
-        private void btnXoa_Click(object sender, EventArgs e)
+        private void dgv_DSSP_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
-            var spXoa = db.SanPhams.SingleOrDefault(sp => sp.MaSp == int.Parse(txtMaMon.Text));
-            DialogResult dialog = MessageBox.Show("Bạn muốn đóng xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-            if (dialog == DialogResult.Yes)
+            if (dgv_DSSP.Columns[e.ColumnIndex].Name == "Xoa")
             {
-                try
+                var spXoa = db.SanPhams.SingleOrDefault(sp => sp.MaSp == int.Parse(txtMaMon.Text));
+                clearTextBox();
+                DialogResult dialog = MessageBox.Show("Bạn muốn đóng xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                if (dialog == DialogResult.Yes)
                 {
-                    db.Remove(spXoa);
-                    db.SaveChanges();
-                    loadDGV();
-                    clearTextBox();
-                }
-                catch (Exception ex)
-                {
-
-                    MessageBox.Show(ex.Message.ToString(), "Thông báo");
+                    try
+                    {
+                        db.Remove(spXoa);
+                        db.SaveChanges();
+                        MessageBox.Show("Đã được xóa!");
+                        loadDGV();
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.Message.ToString(), "Thông báo");
+                    }
                 }
             }
         }
 
-        private void btnFind_Click(object sender, EventArgs e)
+        private void btnRefresh_Click(object sender, EventArgs e)
         {
-            int n = 0;
-            if (int.TryParse(txtFind.Text, out n))
+            loadDGV();
+            txtFind.Clear();
+        }
+
+        private void txtFind_TextChanged(object sender, EventArgs e)
+        {
+            dgv_DSSP.Rows.Clear();
+            try
             {
                 var query = from sp in db.SanPhams
                             join lsp in db.LoaiSanPhams on sp.MaLsp equals lsp.MaLsp
@@ -242,9 +311,15 @@ namespace QLKFC
                                 sp.Loai,
                                 sp.HinhAnh
                             };
-                dgv_DSSP.DataSource = query.ToList();
+                foreach (var item in query)
+                {
+                    if (item.HinhAnh != null)
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia, new Bitmap(pathImage() + item.HinhAnh));
+                    else
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia);
+                }
             }
-            else
+            catch
             {
                 var query = from sp in db.SanPhams
                             join lsp in db.LoaiSanPhams on sp.MaLsp equals lsp.MaLsp
@@ -258,9 +333,63 @@ namespace QLKFC
                                 sp.Loai,
                                 sp.HinhAnh
                             };
-                dgv_DSSP.DataSource = query.ToList();
+                foreach (var item in query)
+                {
+                    if (item.HinhAnh != null)
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia, new Bitmap(pathImage() + item.HinhAnh));
+                    else
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia);
+                }
             }
-            dinhDangBang();
+        }
+
+        private void btnFind_Click(object sender, EventArgs e)
+        {
+            dgv_DSSP.Rows.Clear();
+            try
+            {
+                var query = from sp in db.SanPhams
+                            join lsp in db.LoaiSanPhams on sp.MaLsp equals lsp.MaLsp
+                            where sp.MaSp == int.Parse(txtFind.Text)
+                            select new
+                            {
+                                sp.MaSp,
+                                sp.TenSp,
+                                sp.DonGia,
+                                lsp.TenLsp,
+                                sp.Loai,
+                                sp.HinhAnh
+                            };
+                foreach (var item in query)
+                {
+                    if (item.HinhAnh != null)
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia, new Bitmap(pathImage() + item.HinhAnh));
+                    else
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia);
+                }
+            }
+            catch
+            {
+                var query = from sp in db.SanPhams
+                            join lsp in db.LoaiSanPhams on sp.MaLsp equals lsp.MaLsp
+                            where sp.TenSp.Contains(txtFind.Text)
+                            select new
+                            {
+                                sp.MaSp,
+                                sp.TenSp,
+                                sp.DonGia,
+                                lsp.TenLsp,
+                                sp.Loai,
+                                sp.HinhAnh
+                            };
+                foreach (var item in query)
+                {
+                    if (item.HinhAnh != null)
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia, new Bitmap(pathImage() + item.HinhAnh));
+                    else
+                        dgv_DSSP.Rows.Add(item.MaSp, item.TenSp, item.TenLsp, item.Loai, item.DonGia);
+                }
+            }
         }
         #endregion
     }
